wip: refactor info command
